pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "supervideo"
version = "0.2.0"
description = "Distributed bird classification system for video analysis."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0",
]

[project.optional-dependencies]
classifier = [
    "torch>=2.0",
    "torchvision>=0.15",
    "ultralytics>=8.0",
    "timm>=0.9",
    "numpy>=1.24",
    "opencv-python>=4.8",
]
client = [
    "PySide6>=6.6",
]
dev = [
    "Pillow>=12.1,<13",
]

[project.scripts]
supervideo-frames = "supervideo_frame_extractor.cli:main"
supervideo-client = "client.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
